move console capture to own file
